数据库字段，删除调试信息，新增安卓调试工具

+++ Assets/Sprites/Data/EquipmentData.cs
@@ -12,5 +12,5 @@ public struct EquipmentData
     public int equipment_AP;
     public int equipment_DEF;
     public int equipment_RES;
-    public ulong equipmentPrice;
+    public int equipmentPrice;
 }
